- moves outdated to jobs   - is a function returning an promise   - runs jobs in sequence - adds cronJob in bin #25 - updatedHistory writes "deregistered" date in German format - adds some usage notes as docs to swissmedic and swissmedicHistory
