+++ jobs/swissmedic.js
@@ -65,6 +65,7 @@ function swissmedic(done) {
       log.timeEnd("Swissmedic", "Write Files");
       log.debug("Swissmedic", "Update History");
       log.time("Swissmedic", "Update History");
+      // A build of swissmedic history is strongly coupled with fresh run of this job.
       return swissmedicHistory();
     })
     .then(function () {
