+++ lib/swissmedic/updateHistory.js
@@ -1,6 +1,7 @@
 "use strict";
 
 var merge = require("merge");
+var moment = require("moment");
 
 var objDiff = require("../common/objDiff");
 
@@ -31,7 +32,7 @@ function updateHistory(historyStore, newStore) {
       }
 
       if (!dataInNewStore) {
-        historyStoreData.deregistered = Date.now();
+        historyStoreData.deregistered = moment().format("dd.mm.YYYY");
         deRegistered++;
       }
